requirements.txt
streamlit
chromadb
pandas
google-genai
pysqlite3-binary
